Rename the 'notifications' module to 'events'. Add some TODOs
