+++ src/discord/core.rs
@@ -4,7 +4,7 @@ use std::sync::Arc;
 
 use crate::discord::responses::{BotResponse, NO_RESP};
 
-pub use crate::discord::notifications::GameEvent;
+pub use crate::discord::events::GameEvent;
 
 /// A Subscriber (Module) that listens to the event stream.
 #[async_trait]
@@ -19,7 +19,7 @@ pub trait DiscordSubscriber: Send + Sync {
     }
 
     /// Called for every event. Returns an optional message to send to Discord.
-    async fn on_event(&mut self, event: &crate::discord::notifications::GameEvent, http: &Arc<Http>, channel: ChannelId) -> Vec<BotResponse>;
+    async fn on_event(&mut self, event: &crate::discord::events::GameEvent, http: &Arc<Http>, channel: ChannelId) -> Vec<BotResponse>;
 
     /// Optional periodic task.
     async fn on_tick(&mut self, _http: &Arc<Http>, _channel: ChannelId) -> Vec<BotResponse> {
