+++ src/seh.rs
@@ -101,7 +101,7 @@ unsafe extern "system" fn get_module_base64(hprocess: HANDLE, addr: u64) -> u64 
 
 use winapi::um::winnt::IMAGE_FILE_MACHINE_AMD64;
 
-use crate::discord::notifications::{Crash, GameEvent};
+use crate::discord::events::{Crash, GameEvent};
 fn print_stack(ctx: &mut CONTEXT) -> Vec<String> {
     let process = unsafe { GetCurrentProcess() };
     let thread: HANDLE = HANDLE(-1); // current thread
